Imported reward shaping into vizdoomEnv
Imported the working reward shaping into vizdoomEnv
