+++ A2C/demo.py
@@ -1,6 +1,6 @@
 import os, sys
 sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
-from vizdoomEnv import VizDoomTrain
+from vizdoomEnvPaperDivided import VizDoomTrain
 from stable_baselines3 import DQN, A2C, PPO
 from stable_baselines3.common.monitor import Monitor
 from stable_baselines3.common.evaluation import evaluate_policy
@@ -15,7 +15,7 @@ env = VizDoomTrain(SCENARIO)
 env = Monitor(env)
 #model_dir = fnmatch.filter(os.listdir(DIR), "Trial_" + str(TRIAL_N) + "*")[0]
 
-model = A2C.load("best_model_3000000")
+model = PPO.load("best_model_280000")
 for episode in range(10):
     obs = env.reset()
     done = False
